[INTERNAL][FIX] sap.ui.dt - mutationobserver trigger always root entries
Actually the mutation observer search for relevant parent element of
the mutated dom node. With this change the element that is registered
as root element is called on each mutation related to the root element.
We need to change to this behaviour temporary because of
problems with relative position changes of some elements which do not
trigger mutations. As all overlays are absolut positioned, we need to
recalculate the position of this overlays as well. Due to the missing
mutation for this scenario we just trigger always the root element.

In follow up changes we will adjust the searching algorithm to handle
the case (described above) as well.

Change-Id: Ie4cb76b22c5c2c978a39de147127943b91f88c71
(cherry picked from commit 0341543cd3bf417283c4f3e30d2b0995c4532e2b)
BCP: 2080043811
CR-Id: 002075125900000403772020