[FIX] sap.uxap.ObjectPageLayout: Prevent unwanted snap on scroll
Background:
The height of the bottom spacer of the ObjectPage ensures the page
can be snapped with scroll

Problem:
The two values:
- spacer height
- the snapping breakpoint
are calculated based on obtained DOM metrics, some of which are *rounded*;
On inspection we see:
- spacer value ends up *smaller than needed* [by a fraction of the pixel]
- snapping breakpoint ends up *bigger than needed* [by a fraction of the pixel], as it is ceiled
=> spacer height ends up *smaller than needed* to snap the page with scroll
=> when we scroll to snap,
   the logic in the onscroll listener undesirably expands the page back

Solution:
The first investigated solution was to achieve uniform rounding for both values.
One possibility was found that fixes this particular issue but conflicts
with a previous fix for rounding issue that cannot be retested.

Therefore a solution to suppress the unwanted modification in the
onscroll listener is chosen (as the modification is not really needed,
given the header is already in the desired state before the scroll)

Change-Id: I847dc665c0a27e373d017772450c98ee41fe1ffb
BCP: 1970448827