[INTERNAL] lib/jsdoc: clearly separate missing vs. imported symbols
When gaps in the inheritance tree and interface relationships are closed
with synthetic symbols, they either can be external symbols, known from
the api.json files of predecessor libraries ('import') or they can be
missing, e.g. due to a typo in the reference or due to their visibility
(reference to a private symbols). Only the latter case should be
reported as a warning.

Change-Id: Ide0dd44c849b3b0f0c1e3e0f64dd8c3ee99ed15a