[FIX] sap.m.StepInput: Value State handled correctly + improvements
- Fixes: https://github.com/SAP/openui5/issues/2556

- This change is refactoring of the control, which fixes also
other small issues with the control - some behavior inconsistencies,
value handling and touch issues.

- Now the control have value that is separated from the visual value
displayed, and inner temporary value (because the control value is
not changed on each operation, for example arrow up/down). On each
operation requested, the control calculates the difference between
the temporary value, the value displayed, and the change of value
requested by the operation; then the new calculated value is set as
temporary, and if necessary, as a control value. Now, the value of
the control is set (and the change event is fired) in the following
cases: 1. when increase/decrease buttons are pressed; 2. When Enter
key is pressed and the current value is different than the previous
one; 3. When the control is focused out and the current value is
different than the previous one. In any other cases only the display
value and the temporary values are changed, and change event is not
fired.

Change-Id: I612bba8758ded54ea9de33920e85ebb57eaae87b
BCP: 1970276948
BCP: 1980328638