[FIX] sap.f.ShellBar: Popover not opening on phone is fixed
Problem: The overflow menu of the sap.f.ShellBar does not open on phone,
when the Search is in it. The reason for that is a recent change in the
sap.f.OverflowToolbar which does not render controls with
OverflowToolbarLayoutData = AlwaysOverflow.
Currently, when the Serach is rendered for the very first time (and it
is not opened), it receives LayoutData = Low (which happens in a method,
called each time upon onBeforeRendering). After that, when the
ShellBar executes its _transformToPhoneState method, it sets Always
overflow LayoutData to the Search. When the user tries to open the
Popover menu, the Serach is rendered and in its onBeforeRendering
method, it again sets to itself Low priority LayoutData. This triggers
reseting and invalidation of the sap.f.OverflowToolbar - the Popovers is
getting closed, immediately after it is opened. An error in Console can
be seen.

Solution: In onBeforeEnterOverflow method of the Serach now we call the
same method, which determines which LayoutData should be set, depending
on the logic whether the Search is open or not. There is also one
additional check - whether the Search is in the overflow menu or not.
Because, if it is there, its LayoutData should stay as AlawaysOverflow -
it should not be changed to Low (as it wrongly happened before).

Change-Id: I2c589a468e7b7c0525bd513ef810a243a6c98210
BCP: 2080040770