[INTERNAL] Target.js: Move destruction of component container to component destroy
When routing to a component a ComponentContainer gets created. When
destroying this component the Target class tries to destroy the created
ComponentContainer instance. This was done by intercepting the exit
hook of the component.
When the exit hook is called the Component is not yet fully destroyed.
If the Target destroys the Container the not yet fully destroyed
component gets a second destroy call. This could lead to errors.
Moving the desctruction of the container after the destroy of the
component will fix these issues.

Change-Id: If2f01b6f9fe4f14644dbcb1d293a578823a7d5e4
BCP: 1980539933
CR-Id: 002075125900000482792020