[INTERNAL] Demo Kit: Fixed incorrect RegExp of the Highlighter util
Problem: If we want to highlight terms which consist of
RegExp special characters, we end up with incorrect results.
This happens because if those characters are not escaped, RegExp
recognises them as logical operands / keywords and not literals.

Solution: We escape those special characters by inserting '\'
in front of them in the highlighted term.

Change-Id: Ife5ff1365821ed1a63fbf9b851a045e6bdccf06a
JIRA: BGSOFUIPIRIN-4397