[FIX] Table: Scrolling when inside a ValueHelpDialog
A vertical scroll update process of the scrolling extension might never
finish, because one of it's timers or animation frames might be
cancelled by the table in "onBeforeRendering". In this case, depending
on the priority of the pending process, other processes could not be
started anymore.
The responsibility for the extension's timers and animation frames is
handed over to the extension to avoid this.

Change-Id: I91c8d19b8fb6663aabb290a8415fa7633ac4be15
BCP: 2080018556