[FIX] sap.f.Shellbar: Mega Menu Title doesn`t truncate incorrectly
In current implementation of sap.f.ShellBar we need to add static
width of the sap.m.Button in couple of variants, one of which is
sap.m.MenuButton in MegaMenu, where it cointains title and an
arrow symbol. The width of the arrow + the width of the paddings
need to be added explicitly, so the calculation needs to be correct.

It appears that until now it was wrongly based on Sap Belize theme,
now it is fixed to reflect Sap Quartz metrics. Also there was issue
with rendering of the icon, which now should be fixed.

Change-Id: I4286b89ddffaed2b39777f76bd0fab26479301ce
BCP: 1980478346