[FIX] Demo Kit: Open links in new tab is fixed
Problem: In the static version of the Demo Kit opening links in new tab
does not work.

Solution: Attach to mousedown event and check for auxclick or
contextmenu opened in order to replace the href with correct one and
open it in new tab. In addition, now when click event is fired with ctrl
(for Windows) or metaKey (for Mac) pressed, we open the link in new tab
(as this is a default link behavior).

Change-Id: Ia477a98d446f88c828d6a2de5d09b931e4c0de35
BCP: 2080007465