[INTERNAL][FIX] sap.m.Toolbar: Info Toolbar font is changed
Font italic is no longer used for Info Toolbar. The font is removed from
aggregations: sap.m.Label, sap.m.Text and sap.m.Title, because the text
inside of them was partly cut.
Also the extra spacing, which was supposed to fix the problem is removed,
since it is no longer needed.

Change-Id: I658fa02f2dca6c368c6ce04b2ea35943bda85f4e
BCP: 2080015914