[FIX] NumberFormat: Currency Format Trailing
NumberFormat Currency instances should display the
currency code after to the number (trailing).
Independent from the CLDR pattern.

CLDR json files are adapted such that the new trailing patterns
are part of it.
This increases the size of the CLDR json files, but keeps the
logic in the build time. And the generated patterns can be maintained
more easily.

This is a breaking change with regards to the display of currencies.

Background:
* Consistency: Other SAP UIs show the currency code to the right of the number
  * all SAP UIs should look the same
    * all numbers can be better compared when displayed underneath each other
  * display and edit mode show the same order
    * display mode should show currency code on the right side
    * edit mode shows 2 fields: number field and currency field

Separated Currency tests from NumberFormat tests into separate files
NumberFormat.qunit.js and NumberFormatCurrencies.qunit.js:
* to compare the "trailing" currency from the i18n one
* to reduce the file size of NumberFormat (SOC)

Introduce new Configuration option to have a global switch for this breaking feature.
Test generate with "grunt mochaTest"

BCP: 1970551159
JIRA: CPOUI5FOUNDATION-143
Change-Id: Ic237dffb4f9672de9ee8e70345c26779bd724bd1