[INTERNAL][FIX] Demo Kit: Missing items are now shown in the API Detail
Problem:
Some items from the "sap.ui" namespace which are shown in the
API Reference tree were missing in the API Detail section.

Solution:
We are now adding the missing items in the array which is going
to be displayed in the API Detail section in the Demo Kit's
application logic.

Change-Id: I79af2160e3171a1e2857f245a4f1c84b43658ea6
JIRA: BGSOFUIPIRIN-4044