[INTERNAL] sap.m.Popover: Close behavior Explored sample added
A sample demonstrating how the close behavior of the sap.m.Popover can be controlled.
When the popover is opened, add a style "pointer-events: none" to the area that has
not to be triggered and when the popover is closed, remove it,
e.g. in order to prevent drill-down navigation.

Change-Id: If573d99a06475f426d05edf76c9d5a61830f7436
BCP: BGSOFUIRILA-2693