[INTERNAL][FIX] sap.f.FlexibleColumnLayout: Performance improvement
Problem: Resize handlers of all controls, nested in the first column of
the FlexibleColumnLayout are triggered two times, when navigating from
OneColumnLayout to TwoColumnsMidExpanded layout. They should be triggered
only once with correct width parameter.

Solution: This happens because there is a wrong calculation of Columns'
width, due to a wrong check for visible arrows count. Now the check for
visible arrows is fixed and Columns do not get resized two times.

Change-Id: I2ca4263225d65e84de31741be94d0112bd37dda8
JIRA: BGSOFUIPIRIN-4424