[INTERNAL] sap.m: reduce failures of master-detail/orderbrowser tests
The 'busy indicator' integration tests for both apps fail regularly.

This seems to be caused by the combination of two things:
- to enforce the display of a busy indicator during startup, both apps
  create a mock server that responds with a significant delay
- the integration test then starts the app (with manifest first
  approach) and checks for the busy:true flag
- after metadata has been loaded, the busy flag is unset

But due to the async loading of the app, it can happen that the timer
that implements the mock server's delayed response is processed before
loading of the app has finished. As a result, busy is unset before the
Opa test queue starts checking for busy:true.

With sync loading, the startup was not necessarily faster, but blocked
the main thread and the response could be processed only later.
Additionally, there are effects that might increase the loading time of
the app in the async case (e.g. automatic loading of message bundles).

This change just increases the response delay of the mock server. A
better approach would be to not delay the response by a fixed amount of
time, but to delay it until the app has been initialized + some very
small fixed delayed afterwards.

Change-Id: Ia1f738e5b96a3046b9f9929483f1ce0ef7856637