[FIX] Component Models: Don't create model when dataSource is missing
This fixes a scenario where a model is preloaded and it refers to a
dataSource that is only defined within the parent manifest.

Models are now skipped when no dataSource is found instead of trying to
create them anyways.
Once the component instance is created, the model configuration is read
again with taking the parent manifest into account so that the model can
still be created, but at a later point in time.

Change-Id: Icd1167f7f796666755fa979ffa2eb94abc3615ec
BCP: 0020751295 0000064446 2020