[FEATURE] sap.f.FlexibleColumnLayout: Introduced new property “autoFocus”
Internal NavContainers have autofocus property set to true which
moves the focus when navigating to new columns only once (on render).
We have provided an API to remove the autofocus behavior to comply
with the specifications - when exploring items in a given list,
the focus is expected to stay on the list, so the app user can easily navigate.

Change-Id: If2ebc8a8e8f53e4f40ae931329fccbd3b6cc2499
BCP: 2080069910 2080069750
JIRA: BGSOFUIPIRIN-4486