[INTERNAL][FIX] sap.ui.dt - Stretch test stability improvement
The mutations triggering the applyStyles of the overlays were delayed
through two redundant requestAnimationFrames. It was causing in a
timing issue in some stretch tests. Stretch plugin needs applyStyles to
be finalized before check for stretching is running. Now the
requestAnimationFrame is called just once for a triggered mutation.

Change-Id: I88e682a88938b555f667b2fe0af9b76d840fcf43
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>
BCP: 2080078400