[FIX] sap.uxap.ObjectPageLayout: IconTabBar only visible section show title
Problem: The correct logic when there is only one or only one visible section,
ObjectPageLayout should display the title of this section in the AnchorBar/IconTabBar area.
In the case of IconTabBar and multiple sections but only one section is visible, the title
wasn't displayed.

Solution: In ObjectPageLayout with IconTabBar, we now check for visibility of all the
sections and similar to AnchorBar case we show the title if only one is visible.

Change-Id: Ib8198f7fec798cfdcf5a78efcfe5917b7c389b4d
BCP: 2080090841