[FEATURE] sap.m.FacetFilter: initial data service query optimization option now added
- Problem description:
When a facet popover is opened in order to display a FacetFilterList,
its items aggregation could be fetching long time from the OData model.

- Solution:
By preventing the default behavior and binding the list items inside the
"search" event handler, the internal data fetching is skipped. As a result
no list items are loaded initially. In order to fetch list items a custom
filtering criterion has to be added.

Change-Id: Ide207b576ca3d34e7b3a02a7014e01573f05a61a
JIRA: BGSOFUIBALKAN-5246