[INTERNAL][FIX] AlignedFlowLayout: Unit test false positive fixed
The problem:
Assertion failed error under Firefox ESR 68 and Internet Explorer 11.

The root cause:
An exception is raised when trying to spy on the .disconnect() method
of the ResizeObserver interface, as this method is nonexistent on old
browsers.

The solution:
Bypass the unit test execution of the failing test when the built-in
Resize Observer API is no supported, as there is a fallback for old
browsers.

Change-Id: Ic0d8671eb70e539223821b25ac9fbbb323344bcb