[INTERNAL][FIX] sap.ui.dt - removed debouncing before fire sync event
On objectpage add section the ancorbar buttons were created destroyed
and created again in a row. Due to debouncing in designtime (before
sync event is fired), the triggering cases (create, destroy, create) were
not clearly separated by sync event and got mixed. The mPendingOverlays
map was not cleaned. The second create tried to add a already destroyed
overlay to the parent. The debouncing functionality is now removed.

BCP: 2080116044

Change-Id: Id5bc5379f9a0b9253f461337946612e3599dbfaa
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>