[INTERNAL] sap.m.Slider: Z-index CSS property removed
This change removes all z-index properties from the CSS selectors.
This way the problem mentioned in the incident is resolved.
In order for the Sliders to look normal though, the tickmarks should be
placed before the internal container of the control. This way the
handles and the progress are over the tickmarks as in the visual spec.
The change does not seem to affect the use cases in the following
incidents anymore: 1870534962, 1780348810, for which the z-index was implemented
initially.

Change-Id: Ib9cde22df60b269830bbb3a8e98180bc76857ab1
BCP: 2080117260
CR-Id: 002075125900001097082020