[INTERNAL][FIX] sap.ui.dt - clear mPendingOverlays map on destroy
In the case we have a create overlay (insertAggregation event), destroy
and create again a overlay for the same element we had trouble with the
mPendingOverlays map in DesignTime. On the second create we tried to
add an already destroyed overlay to the parent. Now the mPendingOverlays
map is cleared after the destroy the corresponding overlay.

BCP: 2080116044
CR-Id: 002075125900001118732020

Change-Id: I7ad4396a79f56c70f550e8483e4d8f75b8ea1bdf
Signed-off-by: Arthur Trauter <arthur.trauter@sap.com>