[FIX] sap.m.OverflowToolbar: Issue with infinite rerendering is fixed
Problem: Sometimes, when we have SegmentedButton (ul element with li-s),
it's width is returned by JQuery's outerWidth function differently. It
seems like an edge case, when for example, both li-s are 40.5px and the
ul is 81px. Sometimes, maybe because of internal roundings, the width of
the ul comes as 82px (41 + 41), but other times it comes as 81px (41 +
40).

Solution: Now we use the precise width of the DOM ref of the control,
which is retrieved trough getBoundingClientRect function. When we add the
margins, we ceil the value, because it is better to allocate some more
space for a control than less, as allocating not enough space may result
in triggering new OverflowToolbar's calculations or cutting the control.

Change-Id: Ia0a575f09b64d44498aff6e3cc39f877f10fed75
BCP: 2080039488