[INTERNAL] m.ListBase: improve stability of unit tests
There's a test "BusyIndicator in the middle" which checks whether the
busy indicator is placed within the visible area even for larger scroll
areas. That test expected that the event raised by the internally used
IntersectionObserver is always queued before a setTimeout(...,0).

Obviously, this isn't always true (verified by extensive log tracing).

Simply increasing the delay of the timeout a bit seems to fix the issue.
An alternative solution would be to wait for something more specific,
e.g. for a side effect of the event listener and check the expectations
only then.

Change-Id: Ib0e82fc26436bee002278007fdd0e9f6117a565b