[FIX] sap.m.MultInput: setEditable now recreates the tokens popover
Currently, the n-more popover is created each time the setEditable is called.
This caused an issue when deleting tokens in the tokens popover.
Now, each time the setEditable function is called, the old
list is getting deleted and a new one is created in order to
sync correctly the tokens in the list and inside the tokenizer.

Change-Id: Ibd94cc25c23fa5dcbd4f65f0fdace961dce54d2f
BCP: 2080043629