[FIX] Drag&Drop: prevent D&D indicator from making the body scroll
The D&D Indicator (blue overlay) could cause a scrolling overflow of the
body, so when the user dragged to the edge of the window, the actual UI
could be moved out of the visible area. This only happened in IE+Edge,
but even in Chrome an additional scrollbar was visible.
BCP: 0020751295 0000009139 2020

Change-Id: I68138e9cf45062c740216d3a733f7bdb92feb692