[INTERNAL] sap.uxap.ObjectPageLayout: _updateShowHideState performance
Problem: When the ObjectPageLayout is in Edit scenario, it gets
rerendered when 'Edit' button is clicked, due to new property bindings.
Upon this rerendering all its Sections and SubSections toggle their
visibility, based on their importance property. However, when the DOM
structure is large, this toggling takes too much time. This happens
because jQuery's toggle function, calls getComputedStyle internally,
which forces layout recalculations.

Solution: When there is no change in the current visibility of a
Section/SubSection its visibility is not getting toggled anymore, which
saves a lot of Browser's layout recalculations.

Change-Id: I18dfcea3dcd1dcb1e4c957ee85584415f1eca5bd
JIRA: BGSOFUIPIRIN-4458