[FIX] sap.m.OverflowToolbar: Rounding width value
Problem: When the OverflowToolbar is placed in the main content part of
the DynamicPageTitle its size is determined by its own content (not by
the size of its container). In this case, if the width value of a
control, placed in it, is ceiled, but the Browsers floors or rounds it,
the controls gets in the overflow Popover, despite there might be enough
space for it.

Solution: Round the control width value.

Change-Id: I16bfd9c6ccf4811659aaef8c494aed3e01ef7a44
BCP: 2070005863