[FIX] sap.m.FacetFilter: sap.m.FacetFilterList instance is now visible at all times
Problem description:
sap.m.FacetFilterList visibility was set to "false" when FacetFilter
dialog was opened and navigated to the FacetFilterList page.
This is triggered by a handler which tries to get a sap.m.Bar control
inside the FacetFilterList page. Because of the timing this handler was
called the sap.m.Bar instance was not yet created and inserted into the
sap.m.Page control. The result was taking an element by index from the
sap.m.Page content aggregation, which at that time was the sap.m.FacetFilterList control.

Solution:
When created the sap.m.Bar control, which is representing the all checkbox bar,
is now kept inside an internal property inside sap.m.FacetFilter control instance.
This way we are making sure we are setting the visibility to the right control
if it exists.

Change-Id: Ica2021f12de324a5a5f4d91ba4328b1bf78ccb63
BCP: 2070010203