[INTERNAL] sap.uxap.ObjectPageHeader: performance improvement
Problem: In some cases, the visibility of the ObjectPageHeader's actions
is updated for all actions at the same time, due to data binding, for
example. As each update of an action calls adaptLayout function of the
ObjectPageHeader, this triggers a lot of unneeded layout recalculations
of the Browser, which take too much time.

Solution: The change handler of the actions now is replaced with a new
one, which is delayed. As this method can be called multiple times in
one JavaScript tick, every time it clears the previous timeout, if any,
and sets a new one. This guarantees that the handler will be executed
just once for all simultaneous action buttons' updates.

Change-Id: Iead1dabfb4e445f31ecfc3c955fc0bb1647e7a3e
JIRA: BGSOFUIPIRIN-4459