[INTERNAL] Remove jQuery from LocaleData.js and DateFormat.js
UI5 Web Components have to consume the 1.75 version once released to benefit
from several fixes and changes. This change is mandatory for the UI5 Web Components
to remove the dependency to jQuery. The UI5 Web Components currently use a shim
to mock the jQuery dependency only for mocking a few functions needed by the
DateFormat and the LocaleData. This leads to issues in combination with applications
using jQuery by its own. To avoid those issues in future it is essential to get
rid of the jQuery shim completely in UI5 Web Components.

Change-Id: Ia53b607398581f95774838a1d480d4425124de62
(cherry picked from commit 87d645d9817ac36773cc2851223b825c74a935d4)
CR-Id: 002075125800000191912020
BCP: n/a