[FIX] sap.uxap.ObjectPageSection: Renaming section with new VM is fixed
Problem: When an ObjectPageLayout's Section is renamed, after a new
VariantManagement is created, the first Variant also shows the new name
of the Section, before Browser refresh.

Solution: When setting the revertData now we call
setPropertyBindingOrProperty and when retrieving it, we call
getPropertyBindingOrProperty. The problem came from the fact that when
there is a property binding, if getProperty is called, it returns
undefined. Then, when trying to retrieve the old value and set it to
the Section, we should not check for 'string' type of the revertData,
as it may be a binding path.

Change-Id: Ibfc1a8acb055c65582e4c7a67d42b5d081682b01
BCP: 1970518207