[INTERNAL] Drag&Drop: prevent D&D indicator from making the body scroll
The DnDIndicator (blue overlay) could cause a scrolling overflow of the
body, so when the user dragged to the edge of the window, the actual UI
could be moved out of the visible area. This only happened in IE+Edge,
but even in Chrome an additional scrollbar was visible. A div wrapper
around the DnDIndicator is introduced to fix the issue. The div wrappers
are only added to browers except IE as the div wrapper prevented IE from
scrolling.
BCP: 0020751295 0000009139 2020

Change-Id: I1ef021667b6c58e36c889345160fb5af7b7491db