[FEATURE] sap.f.FlexibleColumnLayout: Restore focus on back navigation
The new property introduced to the FlexibleColumnLayout
is "restoreFocusOnBackNavigation".

Its default value is false.

When the property is set to true, the focus is returned to the last
know focused DOM element after navigating back to a previously
opened column.

For example, closing of the end column and being transferred
back to the mid one.

Change-Id: I0e1ae264d88f74e47c64e8b21bdcba0d9530884d
JIRA: BGSOFUIPIRIN-4454