[INTERNAL] Drag&Drop: prevent D&D indicator from making the body scroll
The DnDIndicator (blue overlay) could cause a scrolling overflow of the
body, so when the user dragged to the edge of the window, the actual UI
could be moved out of the visible area. This only happened in IE+Edge,
but even in Chrome an additional scrollbar was visible. A div wrapper
around the DnDIndicator is introduced to fix the issue. The div wrappers
are only added to browsers except IE as the div wrapper prevented IE
from
scrolling.
BCP: 002075129500000091392020
CR-Id: 002075125800000248452020

Change-Id: Iae7ed80dde9abcc4ce24f6fdf2d0637945bc6444