[INTERNAL] sap.m.MenuButton: remove up-front Menu dependency
sap.m.Menu requires the unified library. And does so in a way that
causes single requests for all modules (no preload).

While the root of this behavior can only be tackled inside Menu, this
change here still has a benefit: any users of MenuButton and Menu can
create the MenuButton synchronously (they might need to), but delay the
Menu creation until the unified lib is loaded properly with
Core.loadLibrary()
One example is the sap.ui.mdc.Table which has exactly this issue right
now.

Change-Id: I35b6b81aa23038f0699f22fbfc3f9148a914522a