[INTERNAL] AlignedFlowLayout: Resize handling optimized (performance)
Optimize resize handling by suppressing the .reflow() method to be
invoked when the height of the layout changes. As currently, there is
no need to reflow the layout when the height changes, but only when
the width changes.

This optimization improves rendering performance reducing unnecessary
recalculation of styles and layouts.

Change-Id: I770e714c5246b317da8f4744d5356e044dba1389