[FIX] sap.m.ComboBox: Filtering now clears previously selectedKey
In some cases, the filtered items may only be shown because of second,
third, etc term matched the typed in by the user value. However, if the ComboBox
has selectedKey already, and this key corresponds to an item, which is already not
visible after the filtering, the selection does not correspond to the users input.
In such cases:
- The selectedKey will be cleared so no "hidden" selection is left in the ComboBox
- Further validation is required from application side as the ComboBox allows input
  that does not match any item from the list.

Change-Id: I07801bcfb51672ba41800365b81b020d23c2a9e8
BCP: 2030117603