[INTERNAL] sap.ui.integration: fix issue with custom values in BooleanEditor & EnumStringEditor
The underlying Combobox component was misused. When the specified value is not from the list
(list of combobox values), e.g. a binding string ("{= ${url}}"), then using just "selectedKey"
property to provide a value to Combobox is not sufficient. On other hand, using just "value" is
also insuffient, because in this scenario if value e.g. "CA" (key) and item title is "California",
it won't be selected automatically. Therefore, both "selectedKey" and "value" should be used
to provide a value to the underlying Combobox control.

Change-Id: I644e43dccf8b0afd7f0f5bb15ad3b8b4fb952878
BCP: 2070026750