[FEATURE] sap.f.cards.AdaptiveContent: Custom theming support
We need to support custom theming for the Adaptive Cards. As they rely
entirely on WebComponents, we need to find a way to proxy classic UI5's
styling/variables into WebComponents.
Later, when someone tries to create a custom theme, would take the base
and would just alter the variables

JIRA: BGSOFUIRILA-2714

Change-Id: I063cc8959f90e4e70f5a67de4c8f63d791169d95