[INTERNAL][FIX] sap.m.HeaderContainerItemNavigator: Suppress dom focus change on mousedown event
Background:
ItemNavigation on mousedown sets dom focus on the one of
the items defined in itemDomRefs array (In this case it is
the container). That triggers onsapfocusleave and interrupts
firing press event on the controls inside of the container
(e.g. delete icon on sap.m.Token in sap.m.MultiComboBox).
In Header container we want the dom focus to stay on the
control inside of the header container on mousedown event.

Fix:
In order to prevent ItemNavigation.focusItem from setting
the focus on the container we temporarily replace
the container's focus function with empty one.

Change-Id: I8cdd699b5d5d1224c388e6a9fcd4ef42fa3aef1b
BCP: 1970549365
Jira: FXUBRQ13-1664