[INTERNAL][FIX] sap.uxap.ObjectPageLayout: Corrected scroll position
Problem:
When navigating through anchor bar buttons, the scroll position is not
correct when the title is hidden.

Root cause:
Height of the header is involved in the calculation of the scroll
position.

Fix:
Since the element is not visible its height is set to 0.

Change-Id: I6a12e1231b30eba3b5bda277edf7add8cd9feffc
BCP: 002075129400002720822020
CR-Id: 002075125800001971792020
(cherry picked from commit d3d68dd355ddbaed57113dd5059c38d51f8b98cc)