[INTERNAL] m.FixFlex (test): do not assert on old DOM after re-rendering
Checking the assertions on DOM that is no longer alive after a
re-rendering (line 232-233) is meaningless for the functionality of the
FixFlex control.

Those assertions only succeeded because of a jQuery feature that falls
back from effective styles (getComputedStyle(node)) to configured styles
(node.style) when DOM is not part of a displayed document (a feature
that was meant for document fragments but also applies to disconnected
nodes in general).

Change-Id: I84c53e74fcb1c981a71b6e91e1b36b9e55441ad8