[INTERNAL] ux3.Shell (test): do not assert on old DOM after re-rendering
Checking the assertions on DOM that is no longer alive after closing a
popup is meaningless for the functionality of the Shell/ToolPopup.

Those assertions only succeeded because of a jQuery feature that falls
back from effective styles (getComputedStyle(node)) to configured styles
(node.style) when DOM is not part of a displayed document (a feature
that was meant for document fragments but also applies to disconnected
nodes in general).

Change-Id: Ibbb10247b04b0c09f381498d8f2df66729664c8a