[FIX] sap.f.DynamicPageTitle: Unneeded truncation of heading is fixed
Problem: On smaller devices the heading of the DynamicPageTitle is
getting truncated while there is enough space for it.

Solution: The problem comes from a specified min-width for all
DynamicPageTitle areas - Heading, Content, Actions) to 48px. Now it is
adjusted to 0, so if there is no content in them - they do not take any
space.
It seems the 48px min-width was specified some time ago, because
of visual degradations. However, this degradations happend because the
min-width of flex context is 'auto' by default. This can make block
elements take up much more space than needed, even stretching their
container beyond the screen edge on small screens. To fix this wrong
behavior it's enough to set the min-width to 0.

Change-Id: I0bcab8083d504be7e5064719926d5a4321d734b2
BCP: 2070041919