[INTERNAL] sap.uxap.AnchorBar: SubSection items are now visibile
Problem:
The hierarchical select of the anchorBar shows
subSections buttons only in the very first renedring.
Once the anchorBar was rerendered, any subSection
buttons were lost

Solution:
Ensured that the lifecycle of items if the hierarchical select
follows the lifecycle of the buttons of the anchorBar's toolbar
(i.e. select items are created when the buttons are created;
  and select items are destroyed when the buttons are destroyed).

Change-Id: If0d9a5bdd0f4d727eeef984c9b444c5ad68f0607