[INTERNAL] sap.ui.integration: get rid of "defaultConfig"
There were no way to replace the "default" configuration of CardEditor.
Using "addDefaultConfig" extended the existent default config. setConfig
was also an extenssion on top of default configuration. As we found out
that the concept of default configuration is not flexible and not
understandable by our consumers, we decided to discontinue it in favor of
simple setConfig() (always replaces everything) and addConfig() (adds the
specified configuration on top of previously set configutation).

Change-Id: I120f9879ab38eeab43f40f269fec6f7ea535df8d
JIRA: UICSFLEX-6185