[INTERNAL] sap.uxap.ObjectPageSubSection: Removed volatile qUnit test
Problem: Test fails on slower machines.

Root Cause: The cached values on which we depend on are from "belize"
theme, but when we do the assert we check "quartz light" metrics.
There is a .25rem difference in a certain padding (between "belize" and
"quartz light"), which leads to 4px difference.
Therefore the test fails with 4px difference.

Solution: Since the test is written a year ago, and it's testing already
removed css styles in the "quartz light" theme, we delete it.

Change-Id: Iead3921dd1499db01384f7d64ed1dbdd4f2670aa
JIRA: BGSOFUIPIRIN-4450
(cherry picked from commit c64d246ab781a41732f70efd50cc9d0490118386)
CR-Id: 002075125800000466612021
BCP: 2180046524