[INTERNAL] Integration Cards: Always use complexParser
So far the card binding parser mode - complex or simple,
was dependent on the application configuration.
Since the cards require complex parser for a lot of features,
and since they handle their binding alone (the applciation is not involved)
the parser for cards is now configured to always be the complex parser.

Change-Id: I49b78f3f5a8686e74471da3fa8525fa8180d3d52