[INTERNAL] sap.f.FlexibleColumnLayout: Inconsistent animation is fixed
Problem: There is an inconsistent behaviour of the FLexibleColumnLayout
when the global animationMode is set to "minimal". In this case, the CSS
transition is not executed, but the JavaScript related code is executed.

Solution: Now the "minimal" setting has the same effect as "none"
setting, which means that the CSS animation is not executed and the
JavaScript related code is also not executed.

Change-Id: Icd98dc8fdb94020a89d616b5dfdcd582b6f78673
JIRA: BGSOFUIPIRIN-4473