[INTERNAL] Table: Fix initial rendering in AutoRowMode
If the table is rendered in a container with a height > 0 and a width = 0, no
rows have been rendered. If the width now changes and the table actually becomes
visible, but the height did not change, the table would not update.
This was caused by the AutoRowMode considering the "offsetWidth" to determine
whether an update is necessary (detect an invisible table). This was change to
consider the "scrollHeight".

Change-Id: Idf03cdbdc1bdb25b16d5ddc317e47336dd95b0c5
CR-Id: 002075125900004175832023
SNOW: CS20230006435634
(cherry picked from commit 270d1cbe06cd550f148c63e1c56cd06834f5de20)