[INTERNAL][FIX] TreeTable: Avoid error when updating the row state
The ODataTreeBinding returns node objects that do not have a context
("{serverIndex: int}") when scrolling to a section for which the data is not
yet available.
If "useGroupMode" is enabled and the "groupHeaderProperty" is set in the
TreeTable, it tries to access methods of the context, which causes an error.

Change-Id: I4e2ce34d49e8499f32530e96ee00c6c7a0d1c572
BCP: 2070087371