[INTERNAL][FIX] sap.uxap.ObjectPageSection: Fixed designtime qUnit
Problem: The designtime qUnit file is some times unstable.

Root Cause: Some times instead of ObjectPageSection, a stashed
"_StashedControl" is passed to a function in the ObjectPageLayout.
This results in unstable code execution in a volatile "if" block.

Solution: We change the "if" block in a way that it's impossible to
have the volatile behavior when a different than the
valid (ObjectPageSection) control is passed.

Change-Id: I6e7b38848730e7547c830460b4a6630856e54826