[INTERNAL][FIX] sap.m.Select: Additional handling for selectedKey
Use case:
Select is used in rows of sap.ui.table.Table. The rows are virtual so
data gets changed during scrolling for each respective row.

Problem:
Once a selectedKey is set, the control has no handling for keys
that corrspond to items that do not exist in the list.

Solution:
Setting a selectedKey that does not exist should bring the control
to its initial state.

BCP: 002028376600000134852020
Change-Id: Iede407f5b09a475dcdaf6a5bf88329e65a2ad240