[INTERNAL] sap.ui.fl: improve handling of changes with open dependencies
when a change is created and applied we check if there are open
dependencies which would lead to the change not being applied after a
reload. In that case previously we didn't even apply it, which led to
inconsistent UIs for move changes. Now the change gets applied and
reverted

Change-Id: I2d18556d6bba48bec751bc130a82129b4e87dff8
JIRA: UICSFLEX-4905