[FIX] Demo Kit: Downloading Samples is fixed
Problem: When you open a Sample which is shown in its own iframe, and
after that you open a new Sample (not shown in iframe) and download it,
index.html and Component.js files are missing.

Solution: Always restore the oModelData iframe's property in
Sample.controller.js, as it is checked in SampleBaseController.js to
determine whether to download the Component.js and index.html files.
This separation of logic is needed becayse iframe examples have a
separate index and component files to describe them.

Change-Id: I0b40ec4e824ca8803f37a7151cbda0ade41d8cb2
BCP: 2070089290