[FIX] sap.f.AvatarGroup: ExploredSamples test now pass in IE11
Problem: When resizing the control, before rendering the "more" button,
the layout was first "breaking", because the last element was on the last place,
but still got its margin-right and the width of element was all elements minus the
one that dropped plus the right margin of the pre-last one. Also jQuery width()
didn`t return the exact control width under IE11,  but appx pixel and a half less. It caused
wrong calculations, leading to infinite loop in all the browsers, optimized by all,
besides IE.

Solution: Rouding the value of width-getting method + CSS fix for the layout now fixes
the issue.

Change-Id: I777216f6ec88904a26e50e5e105949b9276f889e
BCP: 2070068109