[INTERNAL][FIX] sap.f.FlexibleColumnLayout: Fixed volatile behavior
Problem: Focusout event on some of the nav containers is some times
called after the destruction of the FlexibleColumnLayout.
This results in unstable behavior of the FlexibleColumnLayout in some
applications tests.

Solution: We attach function expressions to the nav containers as event
delegates (not anonymous functions as it was), which we detach
on destroy of the FlexibleColumnLayout.

Change-Id: I98b286ccfd49b874ae7afbc1a32adc6ed527a81f
BCP: 2070097015