[FIX] ui.unified.CalendarDateInterval: error not thrown now
Problem description:
When selecting a year range the internally managed focused
date gets updated. After the update the focused date is first
of January. This causes failing internal checks when user
tries to select a year and an error is thrown.

Solution:
The month and date of the internally managed focused date are
kept the same after selecting a year range.

Change-Id: Id270fe462958e80c9a7ce546e356e7cc73619e67